Extract project model to separate crate
