+++ crates/ra_lsp_server/Cargo.toml
@@ -19,7 +19,6 @@ url_serde = "0.2.0"
 lsp-types = "0.55.0"
 walkdir = "2.2.7"
 im = "12.0.0"
-cargo_metadata = "0.7.0"
 rustc-hash = "1.0"
 parking_lot = "0.7.0"
 
@@ -30,6 +29,7 @@ ra_ide_api = { path = "../ra_ide_api" }
 ra_arena = { path = "../ra_arena" }
 gen_lsp_server = { path = "../gen_lsp_server" }
 ra_vfs = { path = "../ra_vfs" }
+ra_project_model = { path = "../ra_project_model" }
 
 [dev-dependencies]
 tempfile = "3"
