+++ crates/ra_syntax/Cargo.toml
@@ -15,7 +15,7 @@ drop_bomb = "0.1.4"
 parking_lot = "0.7.0"
 rowan = "0.3.3"
 
-# ideally, `serde` should be enabled by `ra_lsp_serder`, but we enable it here
+# ideally, `serde` should be enabled by `ra_lsp_server`, but we enable it here
 # to reduce number of compilations
 text_unit = { version = "0.1.6", features = ["serde"] }
 smol_str = { version = "0.1.9", features = ["serde"] }
